backend/scripts/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
python-dateutil>=2.8.0